requirements.txt
Flask
werkzeug==2.3.8
gunicorn
apscheduler
requests
python-dotenv